Chore: refactor 404 (#1547)
