+++ fastify.js
@@ -20,9 +20,7 @@ const {
   kReply,
   kRequest,
   kMiddlewares,
-  kCanSetNotFoundHandler,
-  kFourOhFourLevelInstance,
-  kFourOhFourContext,
+  kFourOhFour,
   kState,
   kOptions,
   kGlobalHooks
@@ -31,6 +29,7 @@ const {
 const { createServer } = require('./lib/server')
 const Reply = require('./lib/reply')
 const Request = require('./lib/request')
+const Context = require('./lib/context')
 const supportedMethods = ['DELETE', 'GET', 'HEAD', 'PATCH', 'POST', 'PUT', 'OPTIONS']
 const buildSchema = require('./lib/validation').build
 const handleRequest = require('./lib/handleRequest')
@@ -43,6 +42,8 @@ const { Schemas, buildSchemas } = require('./lib/schemas')
 const { createLogger } = require('./lib/logger')
 const pluginUtils = require('./lib/pluginUtils')
 const reqIdGenFactory = require('./lib/reqIdGenFactory')
+const build404 = require('./lib/fourOhFour')
+const { beforeHandlerWarning } = require('./lib/warnings')
 const getSecuredInitialConfig = require('./lib/initialConfigValidation')
 const { defaultInitOptions } = getSecuredInitialConfig
 
@@ -74,6 +75,12 @@ function build (options) {
 
   // Instance Fastify components
   const { logger, hasLogger } = createLogger(options)
+
+  // Update the options with the fixed values
+  options.logger = logger
+  options.modifyCoreObjects = modifyCoreObjects
+  options.genReqId = genReqId
+
   // Default router
   const router = FindMyWay({
     defaultRoute: defaultRoute,
@@ -83,7 +90,7 @@ function build (options) {
     versioning: options.versioning
   })
   // 404 router, used for handling encapsulated 404 handlers
-  const fourOhFour = FindMyWay({ defaultRoute: fourOhFourFallBack })
+  const fourOhFour = build404(options)
 
   // HTTP server and its handler
   const httpHandler = router.lookup.bind(router)
@@ -116,9 +123,7 @@ function build (options) {
     [kReply]: Reply.buildReply(Reply),
     [kRequest]: Request.buildRequest(Request),
     [kMiddlewares]: [],
-    [kCanSetNotFoundHandler]: true,
-    [kFourOhFourLevelInstance]: null,
-    [kFourOhFourContext]: null,
+    [kFourOhFour]: fourOhFour,
     [kGlobalHooks]: {
       onRoute: [],
       onRegister: []
@@ -244,7 +249,7 @@ function build (options) {
 
   // Set the default 404 handler
   fastify.setNotFoundHandler()
-  fastify[kFourOhFourLevelInstance] = fastify
+  fourOhFour.arrange404(fastify)
 
   return fastify
 
@@ -534,35 +539,11 @@ function build (options) {
 
         // Must store the 404 Context in 'preReady' because it is only guaranteed to
         // be available after all of the plugins and routes have been loaded.
-        const _404Context = Object.assign({}, this[kFourOhFourContext])
-        _404Context.onSend = context.onSend
-        context[kFourOhFourContext] = _404Context
+        fourOhFour.setContext(this, context)
       })
 
       done(notHandledErr)
     }
-  }
-
-  // Objects that holds the context of every request
-  // Every route holds an instance of this object.
-  function Context (schema, handler, Reply, Request, contentTypeParser, config, errorHandler, bodyLimit, logLevel, attachValidation) {
-    this.schema = schema
-    this.handler = handler
-    this.Reply = Reply
-    this.Request = Request
-    this.contentTypeParser = contentTypeParser
-    this.onRequest = null
-    this.onSend = null
-    this.onError = null
-    this.preHandler = null
-    this.onResponse = null
-    this.config = config
-    this.errorHandler = errorHandler || defaultErrorHandler
-    this._middie = null
-    this._parserOptions = { limit: bodyLimit || null }
-    this.logLevel = logLevel
-    this[kFourOhFourContext] = null
-    this.attachValidation = attachValidation
   }
 
   // HTTP injection handling
@@ -653,132 +634,13 @@ function build (options) {
     if (req.headers['accept-version'] !== undefined) {
       req.headers['accept-version'] = undefined
     }
-    fourOhFour.lookup(req, res)
-  }
-
-  function basic404 (req, reply) {
-    reply.code(404).send(new Error('Not Found'))
-  }
-
-  function fourOhFourFallBack (req, res) {
-    // if this happen, we have a very bad bug
-    // we might want to do some hard debugging
-    // here, let's print out as much info as
-    // we can
-    req.id = genReqId(req)
-    req.originalUrl = req.url
-    var childLogger = logger.child({ reqId: req.id })
-    if (modifyCoreObjects) {
-      req.log = res.log = childLogger
-    }
-
-    childLogger.info({ req }, 'incoming request')
-
-    var request = new Request(null, req, null, req.headers, childLogger)
-    var reply = new Reply(res, { onSend: [], onError: [] }, request, childLogger)
-
-    request.log.warn('the default handler for 404 did not catch this, this is likely a fastify bug, please report it')
-    request.log.warn(fourOhFour.prettyPrint())
-    reply.code(404).send(new Error('Not Found'))
+    fourOhFour.router.lookup(req, res)
   }
 
   function setNotFoundHandler (opts, handler) {
     throwIfAlreadyStarted('Cannot call "setNotFoundHandler" when fastify instance is already started!')
 
-    const _fastify = this
-    const prefix = this[kRoutePrefix] || '/'
-
-    if (this[kCanSetNotFoundHandler] === false) {
-      throw new Error(`Not found handler already set for Fastify instance with prefix: '${prefix}'`)
-    }
-
-    if (typeof opts === 'object') {
-      if (opts.preHandler == null && opts.beforeHandler != null) {
-        beforeHandlerWarning()
-        opts.preHandler = opts.beforeHandler
-      }
-      if (opts.preHandler) {
-        if (Array.isArray(opts.preHandler)) {
-          opts.preHandler = opts.preHandler.map(hook => hook.bind(_fastify))
-        } else {
-          opts.preHandler = opts.preHandler.bind(_fastify)
-        }
-      }
-
-      if (opts.preValidation) {
-        if (Array.isArray(opts.preValidation)) {
-          opts.preValidation = opts.preValidation.map(hook => hook.bind(_fastify))
-        } else {
-          opts.preValidation = opts.preValidation.bind(_fastify)
-        }
-      }
-    }
-
-    if (typeof opts === 'function') {
-      handler = opts
-      opts = undefined
-    }
-    opts = opts || {}
-
-    if (handler) {
-      this[kFourOhFourLevelInstance][kCanSetNotFoundHandler] = false
-      handler = handler.bind(this)
-    } else {
-      handler = basic404
-    }
-
-    this.after((notHandledErr, done) => {
-      _setNotFoundHandler.call(this, prefix, opts, handler)
-      done(notHandledErr)
-    })
-  }
-
-  function _setNotFoundHandler (prefix, opts, handler) {
-    const context = new Context(
-      opts.schema,
-      handler,
-      this[kReply],
-      this[kRequest],
-      this[kContentTypeParser],
-      opts.config || {},
-      this._errorHandler,
-      this[kBodyLimit],
-      this[kLogLevel]
-    )
-
-    avvio.once('preReady', () => {
-      const context = this[kFourOhFourContext]
-
-      const onRequest = this[kHooks].onRequest
-      const preParsing = this[kHooks].preParsing.concat(opts.preParsing || [])
-      const preValidation = this[kHooks].preValidation.concat(opts.preValidation || [])
-      const preSerialization = this[kHooks].preSerialization.concat(opts.preSerialization || [])
-      const preHandler = this[kHooks].preHandler.concat(opts.beforeHandler || opts.preHandler || [])
-      const onSend = this[kHooks].onSend
-      const onError = this[kHooks].onError
-      const onResponse = this[kHooks].onResponse
-
-      context.onRequest = onRequest.length ? onRequest : null
-      context.preParsing = preParsing.length ? preParsing : null
-      context.preValidation = preValidation.length ? preValidation : null
-      context.preSerialization = preSerialization.length ? preSerialization : null
-      context.preHandler = preHandler.length ? preHandler : null
-      context.onSend = onSend.length ? onSend : null
-      context.onError = onError.length ? onError : null
-      context.onResponse = onResponse.length ? onResponse : null
-
-      context._middie = buildMiddie(this[kMiddlewares])
-    })
-
-    if (this[kFourOhFourContext] !== null && prefix === '/') {
-      Object.assign(this[kFourOhFourContext], context) // Replace the default 404 handler
-      return
-    }
-
-    this[kFourOhFourLevelInstance][kFourOhFourContext] = context
-
-    fourOhFour.all(prefix + (prefix.endsWith('/') ? '*' : '/*'), routeHandler, context)
-    fourOhFour.all(prefix || '/', routeHandler, context)
+    fourOhFour.setNotFoundHandler.call(this, opts, handler, avvio, routeHandler, buildMiddie)
   }
 
   // wrapper that we expose to the user for schemas compiler handling
@@ -808,12 +670,6 @@ function build (options) {
     }
 
     return middie
-  }
-
-  function beforeHandlerWarning () {
-    if (beforeHandlerWarning.called) return
-    beforeHandlerWarning.called = true
-    process.emitWarning('The route option `beforeHandler` has been deprecated, use `preHandler` instead')
   }
 }
 
@@ -868,8 +724,7 @@ function override (old, fn, opts) {
   instance[pluginUtils.registeredPlugins] = Object.create(instance[pluginUtils.registeredPlugins])
 
   if (opts.prefix) {
-    instance[kCanSetNotFoundHandler] = true
-    instance[kFourOhFourLevelInstance] = instance
+    instance[kFourOhFour].arrange404(instance)
   }
 
   for (const hook of instance[kGlobalHooks].onRegister) hook.call(this, instance)
@@ -893,16 +748,6 @@ function buildRoutePrefix (instancePrefix, pluginPrefix) {
   }
 
   return instancePrefix + pluginPrefix
-}
-
-function defaultErrorHandler (error, request, reply) {
-  var res = reply.res
-  if (res.statusCode >= 500) {
-    res.log.error({ req: reply.request.raw, res: res, err: error }, error && error.message)
-  } else if (res.statusCode >= 400) {
-    res.log.info({ res: res, err: error }, error && error.message)
-  }
-  reply.send(error)
 }
 
 module.exports = build
