+++ lib/symbols.js
@@ -13,6 +13,7 @@ const keys = {
   kRequest: Symbol('fastify.Request'),
   kMiddlewares: Symbol('fastify.middlewares'),
   kCanSetNotFoundHandler: Symbol('fastify.canSetNotFoundHandler'),
+  kFourOhFour: Symbol('fastify.404'),
   kFourOhFourLevelInstance: Symbol('fastify.404LogLevelInstance'),
   kFourOhFourContext: Symbol('fastify.404ContextKey'),
   kDefaultJsonParse: Symbol('fastify.defaultJSONParse'),
